+++ zinc/src/main/java/xsbti/compile/ZincCompilerUtil.java
@@ -30,7 +30,7 @@ public interface ZincCompilerUtil {
      * @param classpathOptions  The options of all the classpath that the
      *                          compiler takes in.
      * @return A Scala compiler with the given configuration.
-     * @see sbt.inc.ZincBridgeProvider Utility to get the Scala instance using ivy.
+     * @see ZincBridgeProvider Utility to get the Scala instance using ivy.
      */
     public static ScalaCompiler scalaCompiler(ScalaInstance scalaInstance,
                                               File compilerBridgeJar,
@@ -45,7 +45,7 @@ public interface ZincCompilerUtil {
      * @param scalaInstance     The Scala instance to be used.
      * @param compilerBridgeJar The jar file of the compiler bridge.
      * @return A Scala compiler with the given configuration.
-     * @see sbt.inc.ZincBridgeProvider Utility to get the Scala instance using ivy.
+     * @see ZincBridgeProvider Utility to get the Scala instance using ivy.
      */
     public static ScalaCompiler scalaCompiler(ScalaInstance scalaInstance,
                                               File compilerBridgeJar) {
